requirements-backend.txt
# ===============================
# Backend Requirements (Cleaned)
# ===============================

# --- Web API ---
fastapi==0.128.0
uvicorn==0.40.0
starlette==0.50.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
annotated-types==0.7.0
annotated-doc==0.0.4
# CRITICAL ADDITION FOR FILE UPLOADS
python-multipart==0.0.20 

# --- Database ---
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
pgvector==0.4.2
greenlet==3.3.1

# --- Resume Parsing ---
python-docx==1.2.0
pymupdf==1.26.7
beautifulsoup4==4.14.3
lxml==6.0.2
langdetect==1.0.9
unidecode==1.4.0

# --- ML / NLP ---
numpy==1.26.4
pandas==2.3.3
scikit-learn==1.8.0
scipy==1.17.0
torch==2.10.0
sentence-transformers==5.2.2
transformers==5.0.0
tokenizers==0.22.2
huggingface-hub==1.3.4
regex==2026.1.15
joblib==1.5.3

# --- NLP tooling ---
nltk==3.9.2
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
thinc==8.2.4
preshed==3.0.12
cymem==2.0.13
murmurhash==1.0.15
blis==0.7.11
catalogue==2.0.10
srsly==2.5.2
wasabi==1.1.3
weasel==0.4.3
langcodes==3.5.1
cloudpathlib==0.23.0
smart-open==7.5.0
# REMOVED: en-core-web-sm (Handled in Dockerfile)

# --- DVC + S3 ---
dvc[s3]==3.66.1
dvc-data==3.18.2
dvc-objects==5.2.0
dvc-s3==3.3.0
dvc-http==2.32.0
dvc-task==0.40.2
dvc-render==1.0.2
dvc-studio-client==0.22.0
scmrepo==3.6.1
sqltrie==0.11.2
dictdiffer==0.9.0
diskcache==5.6.3
shortuuid==1.0.13
zc-lockfile==4.0

# --- AWS / S3 ---
boto3==1.34.162
botocore==1.34.162
s3fs==2026.1.0
s3transfer==0.10.4
aiobotocore==2.13.3
aiohttp==3.13.3
aiohttp-retry==2.9.1
aioitertools==0.13.0
aiosignal==1.4.0
frozenlist==1.8.0
multidict==6.7.1
yarl==1.22.0

# --- MLflow (Optimized) ---
# REMOVED: mlflow==3.8.1 (Full version brings Flask bloat)
mlflow-skinny==3.8.1
mlflow-tracing==3.8.1
databricks-sdk==0.81.0
protobuf==6.33.4
# REMOVED: flask, flask-cors, waitress, docker, alembic (Not needed for Inference API)

# --- Neo4j ---
neo4j==5.28.3

# --- AI / Reasoning ---
groq==1.0.0
httpx==0.28.1

# --- Utilities ---
requests==2.32.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
tqdm==4.67.1
ruamel-yaml==0.19.1
pyyaml==6.0.3
rich==14.3.1
click==8.3.1
typing-extensions==4.15.0